+++ lib/presentation_layer/components/drawer/nostr_drawer.dart
@@ -1,20 +1,13 @@
-import 'package:camelus/domain_layer/entities/user_metadata.dart';
-
-import 'package:camelus/helpers/nprofile_helper.dart';
-
-import 'package:camelus/presentation_layer/atoms/my_profile_picture.dart';
-
-import 'package:camelus/presentation_layer/providers/metadata_state_provider.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter_svg/flutter_svg.dart';
+
 import 'package:camelus/config/palette.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:package_info_plus/package_info_plus.dart';
-import 'package:qr_flutter/qr_flutter.dart';
-import 'package:url_launcher/url_launcher.dart';
 
+import '../../../domain_layer/entities/user_metadata.dart';
+import '../../atoms/my_profile_picture.dart';
 import '../../providers/following_contact_state_provider.dart';
+import '../../providers/metadata_state_provider.dart';
+import 'nostr_side_menu.dart';
 
 class NostrDrawer extends ConsumerWidget {
   final String pubkey;
@@ -25,63 +18,6 @@ class NostrDrawer extends ConsumerWidget {
 
   void navigateToProfile(BuildContext context) {
     Navigator.pushNamed(context, "/nostr/profile", arguments: pubkey);
-  }
-
-  void _copyToClipboard(BuildContext context, String text) {
-    Clipboard.setData(ClipboardData(text: text));
-    ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-      content: Text("Copied to clipboard: $text"),
-    ));
-  }
-
-  void openQrShareDialog(BuildContext context) async {
-    String nprofile = await NprofileHelper()
-        .getNprofile(pubkey, []); //todo: get recommended relays
-
-    // ignore: use_build_context_synchronously
-    showDialog(
-        context: context,
-        builder: (BuildContext context) {
-          return Dialog(
-            backgroundColor: Palette.extraDarkGray,
-
-            //white border
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(20),
-              //side: const BorderSide(color: Colors.white, width: 1),
-            ),
-            child: Container(
-              padding: const EdgeInsets.all(20),
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  const Text(
-                    "Share your Profile",
-                    style: TextStyle(color: Colors.white),
-                  ),
-                  const SizedBox(height: 40),
-                  QrImageView(
-                    data: "nostr:$nprofile",
-                    version: QrVersions.auto,
-                    size: 300.0,
-                    backgroundColor: Colors.white,
-
-                    //embeddedImage: AssetImage('assets/app_icons/icon.png'),
-                  ),
-                  const SizedBox(height: 20),
-                  GestureDetector(
-                    onTap: () => _copyToClipboard(context, nprofile),
-                    child: Text(
-                      "nostr:$nprofile",
-                      style: const TextStyle(color: Palette.lightGray),
-                    ),
-                  ),
-                  const SizedBox(height: 20),
-                ],
-              ),
-            ),
-          );
-        });
   }
 
   Widget _drawerHeader(context, UserMetadata? metadata, WidgetRef ref) {
@@ -201,171 +137,14 @@ class NostrDrawer extends ConsumerWidget {
     );
   }
 
-  Widget _drawerItem({icon, label, onTap}) {
-    return ListTile(
-      onTap: onTap,
-      leading: SvgPicture.asset(
-        icon,
-        height: 25,
-        colorFilter: const ColorFilter.mode(Palette.gray, BlendMode.srcIn),
-      ),
-      title: Text(label,
-          style: const TextStyle(color: Palette.lightGray, fontSize: 17)),
-    );
-  }
-
-  Widget _textButton({text, onPressed}) {
-    return TextButton(
-        onPressed: onPressed,
-        child: Text(
-          text,
-          style: const TextStyle(color: Palette.extraLightGray, fontSize: 16),
-        ));
-  }
-
-  Widget _divider() {
-    return const Divider(
-      thickness: 0.3,
-      color: Palette.darkGray,
-    );
-  }
-
-  // Add method to get version
-  Future<PackageInfo> _getPackageInfo() async {
-    PackageInfo packageInfo = await PackageInfo.fromPlatform();
-
-    return packageInfo;
-  }
-
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final myUserMetadata =
         ref.watch(metadataStateProvider(pubkey)).userMetadata;
-
     return Drawer(
-      child: Container(
-        color: Palette.background,
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.start,
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            _drawerHeader(context, myUserMetadata, ref),
-            _divider(),
-            _drawerItem(
-                label: 'Profile',
-                icon: 'assets/icons/user.svg',
-                onTap: () {
-                  navigateToProfile(context);
-                }),
-            _drawerItem(
-                label: 'Bookmarks',
-                icon: 'assets/icons/bookmark-simple.svg',
-                onTap: () {
-                  ScaffoldMessenger.of(context).showSnackBar(
-                    const SnackBar(
-                      content: Text('Not implemented yet'),
-                    ),
-                  );
-                }),
-            _drawerItem(
-                label: 'Payments',
-                icon: 'assets/icons/lightning.svg',
-                onTap: () {
-                  ScaffoldMessenger.of(context).showSnackBar(
-                    const SnackBar(
-                      content: Text('Not implemented yet'),
-                    ),
-                  );
-                }),
-            _drawerItem(
-                label: 'Blocklist',
-                icon: 'assets/icons/yin-yang.svg',
-                onTap: () {
-                  // navigate to blocklist
-                  Navigator.pushNamed(context, '/nostr/blockedUsers');
-                }),
-            const Spacer(),
-            const Spacer(),
-            _divider(),
-            Padding(
-                padding: const EdgeInsets.only(left: 20),
-                child: _textButton(
-                    text: 'Settings',
-                    onPressed: () {
-                      Navigator.pushNamed(context, '/settings');
-                    })),
-            const SizedBox(height: 10),
-            Padding(
-                padding: const EdgeInsets.fromLTRB(20, 10, 15, 20),
-                child: _textButton(
-                    text: 'Terms of Service',
-                    onPressed: () {
-                      // lauch url
-                      Uri url = Uri.parse("https://camelus.app/terms");
-                      launchUrl(url, mode: LaunchMode.externalApplication);
-                    })),
-            const Spacer(),
-            Padding(
-              padding: EdgeInsets.only(left: 20),
-              child: FutureBuilder(
-                  future: _getPackageInfo(),
-                  builder: (context, snapshot) {
-                    return Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          'v${snapshot.data?.version}',
-                          style: const TextStyle(
-                            color: Palette.gray,
-                            fontSize: 10,
-                          ),
-                        ),
-                        Text(
-                          'build ${snapshot.data?.buildNumber}',
-                          style: const TextStyle(
-                            color: Palette.gray,
-                            fontSize: 8,
-                          ),
-                        ),
-                        Text(
-                          '${snapshot.data?.buildSignature}',
-                          style: const TextStyle(
-                            color: Palette.gray,
-                            fontSize: 6,
-                          ),
-                        ),
-                      ],
-                    );
-                  }),
-            ),
-            _divider(),
-            Padding(
-              padding: const EdgeInsets.fromLTRB(20, 10, 15, 20),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  SvgPicture.asset(
-                    'assets/icons/sun.svg',
-                    color: Palette.primary,
-                    height: 22,
-                    width: 22,
-                  ),
-                  GestureDetector(
-                    onTap: () {
-                      openQrShareDialog(context);
-                    },
-                    child: SvgPicture.asset(
-                      'assets/icons/qr-code.svg',
-                      color: Palette.primary,
-                      height: 22,
-                      width: 22,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
+      child: NostrSideMenu(
+        leadingWidget: _drawerHeader(context, myUserMetadata, ref),
+        pubkey: pubkey,
       ),
     );
   }
