+++ lib/presentation_layer/layouts/three_colum_layout.dart
@@ -1,3 +1,4 @@
+import 'package:camelus/config/palette.dart';
 import 'package:flutter/material.dart';
 
 class ThreeColumnLayout extends StatelessWidget {
@@ -22,52 +23,55 @@ class ThreeColumnLayout extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return LayoutBuilder(
-      builder: (context, constraints) {
-        final minWidthForRightSidebar =
-            leftSidebarWidth + minMainContentWidth + rightSidebarWidth;
-        final showRightSidebar =
-            constraints.maxWidth >= minWidthForRightSidebar;
+    return Scaffold(
+      backgroundColor: Palette.background,
+      body: LayoutBuilder(
+        builder: (context, constraints) {
+          final minWidthForRightSidebar =
+              leftSidebarWidth + minMainContentWidth + rightSidebarWidth;
+          final showRightSidebar =
+              constraints.maxWidth >= minWidthForRightSidebar;
 
-        // Calculate responsive main content width
-        double mainContentWidth;
-        if (showRightSidebar) {
-          // With right sidebar: use remaining space but respect min/max
-          final availableWidth =
-              constraints.maxWidth - leftSidebarWidth - rightSidebarWidth;
-          mainContentWidth =
-              availableWidth.clamp(minMainContentWidth, maxMainContentWidth);
-        } else {
-          // Without right sidebar: use remaining space but respect min/max
-          final availableWidth = constraints.maxWidth - leftSidebarWidth;
-          mainContentWidth =
-              availableWidth.clamp(minMainContentWidth, maxMainContentWidth);
-        }
+          // Calculate responsive main content width
+          double mainContentWidth;
+          if (showRightSidebar) {
+            // With right sidebar: use remaining space but respect min/max
+            final availableWidth =
+                constraints.maxWidth - leftSidebarWidth - rightSidebarWidth;
+            mainContentWidth =
+                availableWidth.clamp(minMainContentWidth, maxMainContentWidth);
+          } else {
+            // Without right sidebar: use remaining space but respect min/max
+            final availableWidth = constraints.maxWidth - leftSidebarWidth;
+            mainContentWidth =
+                availableWidth.clamp(minMainContentWidth, maxMainContentWidth);
+          }
 
-        final totalWidth = leftSidebarWidth +
-            mainContentWidth +
-            (showRightSidebar ? rightSidebarWidth : 0);
+          final totalWidth = leftSidebarWidth +
+              mainContentWidth +
+              (showRightSidebar ? rightSidebarWidth : 0);
 
-        return Center(
-          child: SizedBox(
-            width: totalWidth,
-            child: Row(
-              children: [
-                SizedBox(width: leftSidebarWidth, child: leftSidebar),
-                SizedBox(
-                  width: mainContentWidth,
-                  child: mainContent,
-                ),
-                if (showRightSidebar)
+          return Center(
+            child: SizedBox(
+              width: totalWidth,
+              child: Row(
+                children: [
+                  SizedBox(width: leftSidebarWidth, child: leftSidebar),
                   SizedBox(
-                    width: rightSidebarWidth,
-                    child: rightSidebar,
+                    width: mainContentWidth,
+                    child: mainContent,
                   ),
-              ],
+                  if (showRightSidebar)
+                    SizedBox(
+                      width: rightSidebarWidth,
+                      child: rightSidebar,
+                    ),
+                ],
+              ),
             ),
-          ),
-        );
-      },
+          );
+        },
+      ),
     );
   }
 }
